Apllying refactorings
